+++ apps/functions-subscription/src/adapters/azure/cosmosdb/trial.ts
@@ -4,8 +4,9 @@ import * as E from 'fp-ts/lib/Either';
 import * as TE from 'fp-ts/lib/TaskEither';
 import * as O from 'fp-ts/lib/Option';
 import { cosmosErrorToDomainError } from './errors';
-import { decodeFromItem, decodeFromFeed } from './decode';
+import { decodeFromFeed, decodeFromItem } from './decode';
 import { TrialCodec, TrialReader, TrialWriter } from '../../../domain/trial';
+import * as RA from 'fp-ts/ReadonlyArray';
 
 const emptyMessageParameter = {
   condition: '',
@@ -20,7 +21,27 @@ export const makeTrialsCosmosContainer = (
   return {
     get: (trialId) =>
       pipe(
-        TE.tryCatch(() => container.item(trialId, trialId).read(), E.toError),
+        TE.tryCatch(
+          () =>
+            container.items
+              .query({
+                query: 'SELECT * FROM c WHERE c.id = @id OFFSET 0 LIMIT 1',
+                parameters: [
+                  {
+                    name: '@id',
+                    value: trialId,
+                  },
+                ],
+              })
+              .fetchAll(),
+          E.toError,
+        ),
+        TE.flatMapEither(decodeFromFeed(TrialCodec)),
+        TE.mapBoth(cosmosErrorToDomainError, RA.head),
+      ),
+    getByIdAndOwnerId: (trialId, ownerId) =>
+      pipe(
+        TE.tryCatch(() => container.item(trialId, ownerId).read(), E.toError),
         TE.flatMapEither(decodeFromItem(TrialCodec)),
       ),
     list: (pageSize, maximumId?, minimumId?) =>