+++ apps/functions-subscription/src/domain/trial.ts
@@ -6,7 +6,7 @@ import { pipe } from 'fp-ts/function';
 import * as RTE from 'fp-ts/ReaderTaskEither';
 import { Capabilities } from './capabilities';
 import { ItemAlreadyExists } from './errors';
-import { User } from './users';
+import { Tenant, TenantIdCodec } from './users';
 
 // a unique brand for trialId
 interface TrialIdBrand {
@@ -21,8 +21,8 @@ export const TrialIdCodec = t.brand(
 export type TrialId = t.TypeOf<typeof TrialIdCodec>;
 
 const BaseTrialCodec = t.intersection([
-  t.strict({ id: TrialIdCodec, name: NonEmptyString }),
-  t.partial({ description: t.string, ownerId: t.string }), //ownerId should be mandatory after the migration
+  t.strict({ id: TrialIdCodec, name: NonEmptyString, ownerId: TenantIdCodec }),
+  t.partial({ description: t.string }), //ownerId should be mandatory after the migration
 ]);
 
 const CreatingTrialCodec = t.intersection([
@@ -57,6 +57,10 @@ export interface TrialReader {
     maximumId?: TrialId | null,
     minimumId?: TrialId | null,
   ) => TE.TaskEither<Error, readonly Trial[]>;
+  readonly getByIdAndOwnerId: (
+    trialId: TrialId,
+    ownerId: Trial['ownerId'],
+  ) => TE.TaskEither<Error, O.Option<Trial>>;
 }
 
 const makeTrialId = () =>
@@ -69,7 +73,7 @@ const makeTrialId = () =>
 const makeTrial = (
   name: Trial['name'],
   description: Trial['description'],
-  ownerId: User['id'],
+  ownerId: Tenant['id'],
 ) =>
   pipe(
     makeTrialId(),
@@ -85,7 +89,7 @@ const makeTrial = (
 export const insertTrial = (
   name: Trial['name'],
   description: Trial['description'],
-  { id: ownerId }: User,
+  { id: ownerId }: Tenant,
 ) =>
   pipe(
     RTE.ask<Pick<Capabilities, 'trialWriter'>>(),
@@ -95,10 +99,14 @@ export const insertTrial = (
     ),
   );
 
-export const getTrialById = (trialId: TrialId) =>
+export const getTrialById = (trialId: TrialId, tenant: Tenant) =>
   pipe(
     RTE.ask<Pick<Capabilities, 'trialReader'>>(),
-    RTE.flatMapTaskEither(({ trialReader }) => trialReader.get(trialId)),
+    RTE.flatMapTaskEither(({ trialReader }) =>
+      tenant.type === 'owner'
+        ? trialReader.getByIdAndOwnerId(trialId, tenant.id)
+        : trialReader.get(trialId),
+    ),
   );
 
 export const listTrials = (